major overhaul of event.cpp
