+++ projects/rocprofiler-sdk/source/lib/rocprofiler-sdk/hip/event.hpp
@@ -22,19 +22,13 @@
 
 #pragma once
 
-#include "lib/common/container/small_vector.hpp"
-#include "lib/common/synchronized.hpp"
+#include "lib/rocprofiler-sdk/hsa/rocprofiler_packet.hpp"
 #include "lib/rocprofiler-sdk/tracing/fwd.hpp"
-#include "lib/rocprofiler-sdk/tracing/profiling_time.hpp"
 
-#include <memory>
-
-#include <rocprofiler-sdk/callback_tracing.h>
 #include <rocprofiler-sdk/fwd.h>
 
-#include <hip/hip_runtime_api.h>
-
 #include <cstdint>
+#include <optional>
 #include <vector>
 
 namespace rocprofiler
@@ -43,6 +37,10 @@ namespace context
 {
 struct correlation_id;
 }
+namespace hsa
+{
+class Queue;
+}
 namespace hip
 {
 namespace event
@@ -50,111 +48,82 @@ namespace event
 const char*
 name_by_id(uint32_t id);
 
-uint32_t
-id_by_name(const char* name);
-
-std::vector<const char*>
-get_names();
-
 std::vector<uint32_t>
 get_ids();
-
-using profiling_time = tracing::profiling_time;
-
-void
-barrier_complete(tracing::tracing_data&                        tracing_data_v,
-                 rocprofiler_thread_id_t                       tid,
-                 uint64_t                                      internal_corr_id,
-                 uint64_t                                      ancestor_corr_id,
-                 profiling_time                                barrier_time,
-                 rocprofiler_hip_event_operation_t             operation,
-                 rocprofiler_callback_tracing_hip_event_data_t callback_record);
-
-struct active_event_context_t
-{
-    rocprofiler_hip_event_operation_t operation        = ROCPROFILER_HIP_EVENT_NONE;
-    uint64_t                          hip_event_handle = 0;
-    bool                              barrier_captured = false;
-};
-
-active_event_context_t*
-get_active_event_context();
-
-struct event_record_info_t
-{
-    rocprofiler_queue_id_t queue_id        = {.handle = 0};
-    rocprofiler_agent_id_t agent_id        = {.handle = 0};
-    uint64_t               original_signal = 0;
-    // Set when the record barrier for this generation completes on the GPU. Reset by
-    // record_event_info when the event is re-recorded.
-    bool completed = false;
-};
-
-void
-record_event_info(uint64_t hip_event_handle, event_record_info_t info);
-
-// Marks the current record generation as complete. CLR short-circuits waits on a
-// completed event, so register_deferred_wait stops tracking them once this is set.
-void
-mark_event_completed(uint64_t hip_event_handle);
-
-event_record_info_t
-lookup_event_info(uint64_t hip_event_handle);
-
-struct coalesce_pending_t
-{
-    tracing::tracing_data                         tracing_data     = {};
-    rocprofiler_callback_tracing_hip_event_data_t callback_record  = {};
-    rocprofiler_thread_id_t                       tid              = 0;
-    uint64_t                                      internal_corr_id = 0;
-    uint64_t                                      ancestor_corr_id = 0;
-    context::correlation_id*                      corr_id_ref      = nullptr;
-};
-
-struct coalesce_group_t
-{
-    profiling_time                                         barrier_time = {};
-    bool                                                   completed    = false;
-    common::container::small_vector<coalesce_pending_t, 4> pending      = {};
-};
-
-using coalesce_group_ptr_t = std::shared_ptr<common::Synchronized<coalesce_group_t>>;
-
-void
-store_coalesce_group(uint64_t hip_event_handle, coalesce_group_ptr_t group);
-
-coalesce_group_ptr_t
-lookup_coalesce_group(uint64_t hip_event_handle);
-
-void
-erase_event_info(uint64_t hip_event_handle);
-
-struct pending_wait_t
-{
-    tracing::tracing_data    tracing_data     = {};
-    rocprofiler_thread_id_t  tid              = 0;
-    uint64_t                 internal_corr_id = 0;
-    uint64_t                 ancestor_corr_id = 0;
-    context::correlation_id* corr_id_ref      = nullptr;
-    uint64_t                 hip_event_handle = 0;
-    event_record_info_t      source_info      = {};
-};
-
-void
-register_pending_wait(uint64_t signal_handle, pending_wait_t pw);
-
-bool
-has_pending_waits();
-
-pending_wait_t
-consume_pending_wait(uint64_t signal_handle);
-
-pending_wait_t
-consume_pending_wait_for_handle(uint64_t signal_handle, uint64_t hip_event_handle);
 
 template <typename TableT>
 void
 update_table(TableT* table);
+
+// Single gate for the queue interceptor fast paths: true when a HIP event API call is
+// active on this thread or any deferred wait is outstanding. Independent of the gates for
+// other tracing features, which remain separately testable.
+bool
+is_active();
+
+// What the queue interceptor should emit in place of a candidate barrier packet.
+// When intercepted is false the caller emits the original packet unchanged.
+struct barrier_plan_t
+{
+    bool                    intercepted    = false;
+    hsa::rocprofiler_packet barrier        = {};
+    bool                    has_forwarding = false;
+    hsa::rocprofiler_packet forwarding     = {};
+};
+
+// Decides whether pkt is the GPU barrier produced by a hipEventRecord or
+// hipStreamWaitEvent on this thread. When it is, substitutes a pooled completion signal,
+// fires the ENTER/EXIT callbacks, and registers the handler that emits the record. All
+// event bookkeeping stays inside this call; the caller only emits packets.
+//
+// is_barrier is supplied by the caller because AQL packet classification (including the
+// vendor BARRIER_VALUE header decode) belongs to the queue layer, and duplicating it here
+// would risk the two copies drifting apart.
+barrier_plan_t
+plan_barrier(hsa::Queue&                    queue,
+             const hsa::rocprofiler_packet& pkt,
+             bool                           is_barrier,
+             const tracing::tracing_data&   tracing_data_v,
+             context::correlation_id*       corr_id);
+
+// A hipStreamWaitEvent's GPU dependency is often folded by CLR into a later packet's
+// barrier rather than emitted as a standalone one. The interceptor reports candidate
+// dependency signals here as it walks a packet batch; any wait registered against one is
+// claimed and staged.
+//
+// Staging is thread-local and scoped to a single packet batch: every claim is followed by
+// either bind_staged_waits or flush_staged_waits before the interceptor returns, so
+// nothing carries over into the next batch on this thread.
+void
+claim_deferred_wait(uint64_t dep_signal_handle);
+
+// Claims deferred waits carried as dependencies of a barrier packet, reading whichever
+// dependency field matches the packet type. No-op when this barrier was itself the one
+// intercepted for a record or wait, since that path reports its own completion.
+//
+// Must be called after plan_barrier for the same packet: the two share the per-thread
+// record of whether that barrier was already captured.
+void
+claim_deferred_waits(const hsa::rocprofiler_packet& pkt, uint32_t packet_type, bool is_barrier);
+
+// Attach the staged waits to a submission that already has a completion signal. The
+// caller must invoke submission_complete for that signal before returning it to the
+// signal pool, otherwise the pool can hand the same handle to another submission while
+// waits are still bound to it.
+void
+bind_staged_waits(hsa_signal_t completion_signal);
+
+// Used when a batch has staged waits but no kernel packet whose signal they can ride on.
+// Allocates a signal, registers a private completion handler, and returns the barrier
+// packet the caller should append. Returns nullopt when nothing is staged.
+std::optional<hsa::rocprofiler_packet>
+flush_staged_waits(hsa::Queue& queue, rocprofiler_timestamp_t enqueue_ts);
+
+// Emit a WAIT record for every wait bound to this completion signal.
+void
+submission_complete(const hsa::Queue&       queue,
+                    hsa_signal_t            completion_signal,
+                    rocprofiler_timestamp_t enqueue_ts);
 
 }  // namespace event
 }  // namespace hip
